backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
python-dotenv==1.0.0
argon2-cffi==21.3.0
PyJWT==2.8.0
pydantic[email]==1.10.7
alembic==1.11.1
requests==2.31.0
redis==4.5.4
email-validator==1.3.1
pytest==7.4.0
